Add --print-ifname option to return host facing interface name used
A script may need to know about the actual host facing interface that
was created for the bridged use case to further manipulate the port.

Adds a new optional --print-ifname option to echo the host facing
interface name. Also adds an optional --quiet option to avoid printing
warnings to make --print-ifname reliable. --quiet is implied if
--print-ifname is set.

Signed-off-by: Thomas Graf <tgraf@noironetworks.com>